+++ backend/src/main/java/ca/metricalsky/winston/repository/VideoRepository.java
@@ -2,7 +2,6 @@ package ca.metricalsky.winston.repository;
 
 import ca.metricalsky.winston.entity.Video;
 import ca.metricalsky.winston.entity.view.VideoCount;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
@@ -14,11 +13,6 @@ import java.util.Optional;
 @Repository
 public interface VideoRepository extends JpaRepository<Video, String> {
 
-    List<Video> findAllByChannelIdOrderByPublishedAtDesc(String channelId);
-
-    List<Video> findByChannelIdOrderByPublishedAtDesc(String channelId, Pageable pageable);
-
-
     @Query("""
             SELECT c.id AS channelId, COUNT(v.id) AS videos
             FROM Channel c
@@ -26,6 +20,13 @@ public interface VideoRepository extends JpaRepository<Video, String> {
             GROUP BY c.id
             """)
     List<VideoCount> countAllByChannelId();
+
+    @Query("""
+            SELECT v FROM Video v
+            WHERE v.channelId = (SELECT id FROM Channel WHERE customUrl = :channelHandle)
+            ORDER BY v.publishedAt DESC
+            """)
+    List<Video> findAllByChannelHandle(String channelHandle);
 
     @Query("SELECT MAX(v.publishedAt) FROM Video v WHERE v.channelId = :channelId")
     Optional<OffsetDateTime> findLastPublishedAtForChannelId(String channelId);