+++ backend/src/main/java/ca/metricalsky/winston/service/AuthorService.java
@@ -5,8 +5,11 @@ import ca.metricalsky.winston.mapper.dto.AuthorDtoMapper;
 import ca.metricalsky.winston.repository.AuthorRepository;
 import lombok.RequiredArgsConstructor;
 import org.mapstruct.factory.Mappers;
+import org.springframework.data.util.Optionals;
 import org.springframework.stereotype.Service;
 
+import java.net.URLEncoder;
+import java.nio.charset.StandardCharsets;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Optional;
@@ -14,6 +17,8 @@ import java.util.Optional;
 @Service
 @RequiredArgsConstructor
 public class AuthorService {
+
+    private static final String CHANNEL_URL_PREFIX = "http://www.youtube.com/c/";
 
     private final AuthorDtoMapper authorDtoMapper = Mappers.getMapper(AuthorDtoMapper.class);
     private final AuthorRepository authorRepository;
@@ -26,8 +31,15 @@ public class AuthorService {
                 .toList();
     }
 
-    public Optional<AuthorDto> findById(String authorId) {
-        return authorRepository.findById(authorId)
-                .map(authorDtoMapper::toAuthorDto);
+    public Optional<AuthorDto> findByHandle(String authorHandle) {
+        return Optionals.firstNonEmpty(
+                () -> authorRepository.findByDisplayName(authorHandle),
+                () -> authorRepository.findByChannelUrl(getChannelUrl(authorHandle)),
+                () -> authorRepository.findById(authorHandle)
+        ).map(authorDtoMapper::toAuthorDto);
+    }
+
+    private static String getChannelUrl(String authorHandle) {
+        return CHANNEL_URL_PREFIX + URLEncoder.encode(authorHandle, StandardCharsets.UTF_8);
     }
 }