+++ backend/src/main/java/ca/metricalsky/winston/web/VideoController.java
@@ -22,15 +22,15 @@ public class VideoController {
         this.channelService = channelService;
     }
 
-    @GetMapping("/api/channels/{channelId}/videos")
-    public List<VideoDto> listByChannelId(@PathVariable String channelId) {
-        return videoService.findAllByChannelId(channelId);
+    @GetMapping("/api/channels/{channelHandle}/videos")
+    public List<VideoDto> listByChannelHandle(@PathVariable String channelHandle) {
+        return videoService.findAllByChannelHandle(channelHandle);
     }
 
     @GetMapping("/api/videos/{videoId}")
     public VideoDto findById(@PathVariable String videoId) {
         var video = videoService.getById(videoId);
-        channelService.findById(video.getChannelId()).ifPresent(channel -> {
+        channelService.findByHandle(video.getChannelId()).ifPresent(channel -> {
             video.setChannel(channel);
             video.setChannelId(null);
         });