+++ frontend/src/routes/channels/id/VideoCard.tsx
@@ -4,6 +4,7 @@ import { Link } from "react-router";
 import { ArrowUpLeftCircleFill } from "react-bootstrap-icons";
 import { Date } from "../../../components/Date";
 import { CommentCounts } from "../../../components/comments/CommentCounts";
+import { routes } from "../../../utils/links";
 
 type VideoCardProps = {
     video: VideoWithChannelIdDto;
@@ -13,13 +14,13 @@ export const VideoCard = ({ video }: VideoCardProps) => (
     <Col className={"g-2"}>
         <Card className={"h-100"}>
             <Ratio aspectRatio={"4x3"} className={"bg-secondary-subtle"}>
-                <Link to={`/videos/${video.id}`}>
+                <Link to={routes.videos.details(video)}>
                     <Card.Img variant={"top"} src={video.thumbnailUrl} />
                 </Link>
             </Ratio>
             <Card.Body>
                 <Card.Subtitle>
-                    <Link to={`/videos/${video.id}`}>{video.title}</Link>
+                    <Link to={routes.videos.details(video)}>{video.title}</Link>
                 </Card.Subtitle>
             </Card.Body>
             <Card.Footer>